Renaming the example to p2p-webrtc
