+++ examples/p2p-webrtc/video-transform/client/typescript/vite.config.js
