+++ examples/p2p-webrtc/voice-agent/index.html
